refactor: library/map 폴더 정리
