+++ src/library/map/atoms/searchResult.ts
