+++ src/library/map/utils/createWindowContent.ts
