+++ src/library/map/utils/generateMarker.ts
